requirements.txt
Pillow >= 9.0.0
requests >= 1.9.0